+++ AintBnB.WebApi/Controllers/BookingController.cs
@@ -1,11 +1,16 @@
-﻿using AintBnB.BusinessLogic.Interfaces;
+﻿using AintBnB.BusinessLogic.CustomExceptions;
+using AintBnB.BusinessLogic.Interfaces;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using System;
 using System.Threading.Tasks;
+using static AintBnB.BusinessLogic.Helpers.Authentication;
+using static AintBnB.WebApi.Helpers.CurrentUserDetails;
 
 namespace AintBnB.WebApi.Controllers
 {
     [ApiController]
+    [Authorize]
     public class BookingController : ControllerBase
     {
         private IBookingService _bookingService;
@@ -21,22 +26,26 @@ namespace AintBnB.WebApi.Controllers
             _deletionService = deletionService;
         }
 
-        /// <summary>API GET request to make a booking.</summary>
+        /// <summary>API POST request to make a booking.</summary>
         /// <param name="startDate">The start date of the booking.</param>
         /// <param name="bookerId">The user-ID of the booker.</param>
         /// <param name="nights">The amount of nights to book.</param>
         /// <param name="accommodationId">The ID of the accommodation to book.</param>
         /// <returns>Status 200 and the booking if successful, otherwise status code 400</returns>
-        [HttpGet]
-        [Route("api/[controller]/{startDate}/{bookerId}/{nights}/{accommodationId}")]
-        public async Task<IActionResult> BookAsync([FromRoute] string startDate, [FromRoute] int bookerId, [FromRoute] int nights, [FromRoute] int accommodationId)
+        [HttpPost]
+        [Route("api/[controller]/book")]
+        public async Task<IActionResult> BookAsync([FromBody] string[] bookingInfo)
         {
             try
             {
-                var booker = await _userService.GetUserAsync(bookerId);
-                var accommodation = await _accommodationService.GetAccommodationAsync(accommodationId);
-                var booking = await _bookingService.BookAsync(startDate, booker, nights, accommodation);
-                return Ok(booking);
+                var accommodation = await _accommodationService.GetAccommodationAsync(int.Parse(bookingInfo[3]));
+                var booker = await _userService.GetUserAsync(int.Parse(bookingInfo[1]));
+
+                if (!CheckIfUserIsAllowedToPerformAction(booker, GetIdOfLoggedInUser(HttpContext), GetUsertypeOfLoggedInUser(HttpContext)))
+                    return BadRequest($"Must be performed by a customer with ID {booker.Id}, or by admin on behalf of a customer with ID {booker.Id}!");
+
+                var booking = await _bookingService.BookAsync(bookingInfo[0], booker, int.Parse(bookingInfo[2]), accommodation);
+                return Created(HttpContext.Request.Scheme + "://" + HttpContext.Request.Host + HttpContext.Request.Path + "/" + booking.Id, booking);
             }
             catch (Exception ex)
             {
@@ -44,20 +53,24 @@ namespace AintBnB.WebApi.Controllers
             }
         }
 
-        /// <summary>API GET request to update an existing booking.</summary>
+        /// <summary>API PUT request to update an existing booking.</summary>
         /// <param name="newStartDate">The updated start date of the booking.</param>
         /// <param name="nights">The amount of nights to book.</param>
         /// <param name="bookingId">The ID of the booking to update.</param>
         /// <returns>Status 200 and the updated booking if successful, otherwise status code 400</returns>
-        [HttpGet]
-        [Route("api/[controller]/{newStartDate}/{nights}/{bookingId}")]
-        public async Task<IActionResult> UpdateBookingAsync([FromRoute] string newStartDate, [FromRoute] int nights, [FromRoute] int bookingId)
+        [HttpPut]
+        [Route("api/[controller]/{bookingId}")]
+        public async Task<IActionResult> UpdateBookingAsync([FromBody] string[] newDates, [FromRoute] int bookingId)
         {
+            var booking = await _bookingService.GetBookingAsync(bookingId);
+
+            if (!CheckIfUserIsAllowedToPerformAction(booking.BookedBy, GetIdOfLoggedInUser(HttpContext), GetUsertypeOfLoggedInUser(HttpContext)))
+                return BadRequest($"Must be performed by the booker, or by admin on behalf of the booker!");
+
             try
             {
-                var booking = await _bookingService.GetBookingAsync(bookingId);
-                await _bookingService.UpdateBookingAsync(newStartDate, nights, bookingId);
-                return Ok(booking);
+                await _bookingService.UpdateBookingAsync(newDates[0], int.Parse(newDates[1]), bookingId);
+                return Ok();
             }
             catch (Exception ex)
             {
@@ -65,18 +78,23 @@ namespace AintBnB.WebApi.Controllers
             }
         }
 
-        /// <summary>API GET request to leave a rating on a booking</summary>
+        /// <summary>API POST request to leave a rating on a booking</summary>
         /// <param name="bookingId">The ID of the booking to leave a rating for</param>
         /// <param name="rating">The rating from 1-5</param>
         /// <returns>Status 200 if successful, otherwise status code 400</returns>
-        [HttpGet]
-        [Route("api/[controller]/rate/{bookingId}/{rating}")]
-        public async Task<IActionResult> LeaveRatingAsync([FromRoute] int bookingId, [FromRoute] int rating)
+        [HttpPost]
+        [Route("api/[controller]/rate/{bookingId}")]
+        public async Task<IActionResult> LeaveRatingAsync([FromRoute] int bookingId, [FromBody] int rating)
         {
+            var booker = await _bookingService.GetBookingAsync(bookingId);
+
+            if (booker.BookedBy.Id != GetIdOfLoggedInUser(HttpContext))
+                return BadRequest(new AccessException("Only the booker can leave a rating!").Message);
+
             try
             {
                 await _bookingService.RateAsync(bookingId, rating);
-                return Ok("Rated");
+                return Ok();
             }
             catch (Exception ex)
             {
@@ -93,7 +111,12 @@ namespace AintBnB.WebApi.Controllers
         {
             try
             {
-                return Ok(await _bookingService.GetBookingAsync(id));
+                var booking = await _bookingService.GetBookingAsync(id);
+
+                if (CorrectUserOrOwnerOrAdmin(booking.Accommodation.Owner.Id, booking.BookedBy.Id, GetIdOfLoggedInUser(HttpContext), GetUsertypeOfLoggedInUser(HttpContext)))
+                    return Ok(booking);
+                else
+                    return NotFound(new AccessException().Message);
             }
             catch (Exception ex)
             {
@@ -108,14 +131,21 @@ namespace AintBnB.WebApi.Controllers
         [Route("api/[controller]/{id}/bookingsownaccommodation")]
         public async Task<IActionResult> GetBookingsOnOwnedAccommodationsAsync([FromRoute] int id)
         {
-            try
+            var user = await _userService.GetUserAsync(id);
+
+            if (CorrectUserOrAdmin(user.Id, GetIdOfLoggedInUser(HttpContext), GetUsertypeOfLoggedInUser(HttpContext)))
             {
-                return Ok(await _bookingService.GetBookingsOfOwnedAccommodationAsync(id));
+                try
+                {
+                    return Ok(await _bookingService.GetBookingsOfOwnedAccommodationAsync(id));
+                }
+                catch (Exception ex)
+                {
+                    return NotFound(ex.Message);
+                }
             }
-            catch (Exception ex)
-            {
-                return NotFound(ex.Message);
-            }
+            else
+                return NotFound(new AccessException().Message);
         }
 
         /// <summary>API GET request to return all the bookings from the database</summary>
@@ -126,7 +156,12 @@ namespace AintBnB.WebApi.Controllers
         {
             try
             {
-                return Ok(await _bookingService.GetAllBookingsAsync());
+                var user = await _userService.GetUserAsync(GetIdOfLoggedInUser(HttpContext));
+
+                if (AdminChecker(user.UserType))
+                    return Ok(await _bookingService.GetAllInSystemAsync());
+                else
+                    return Ok(await _bookingService.GetOnlyOnesOwnedByUserAsync(GetIdOfLoggedInUser(HttpContext)));
             }
             catch (Exception ex)
             {
@@ -143,8 +178,14 @@ namespace AintBnB.WebApi.Controllers
         {
             try
             {
+                var booking = await _bookingService.GetBookingAsync(id);
+
+                if (!CorrectUserOrOwnerOrAdmin(booking.Accommodation.Owner.Id, booking.BookedBy.Id, GetIdOfLoggedInUser(HttpContext), GetUsertypeOfLoggedInUser(HttpContext)))
+                    return BadRequest(new AccessException().Message);
+
                 await _deletionService.DeleteBookingAsync(id);
-                return Ok("Deletion ok");
+
+                return Ok();
             }
             catch (Exception ex)
             {
